refactor(analysis): split analyze-source-snapshots into focused modules
Extract analyze-source-snapshots.ts (995 lines) into:
- import-graph-analyzer.ts: regex-based import/export/require parsing
- dependency-graph-builder.ts: file dependency graph construction
- architecture-context-merger.ts: architecture context integration
- analyze-source-snapshots.ts: thin orchestrator (246 lines)

Closes #172
